Fix split in decodeFilter()
